Cleaning up bloat.
